Extracted the metadata helper functions
